Pull location text header into own component
